icept: initial client/server test
Preparation for an interception test.

Create a simple client/server pair that exchange 1 byte messages.
The client sends message 'a' to the server, which responds with
message 'b'.

Run the test both over IPv4 and IPv6.

The processes run in private network namespaces and communicate
over veth. This state is setup with helper script with_veth.sh.

Signed-off-by: Willem de Bruijn <willemb@google.com>