ktls: example server
Create a simple TCP server that echoes client input. Run over TLS and
optionally off-load the datapath to kTLS.

To validate the implementation, communicate with a standard client:
either socat or openssl s_client.

In kTLS mode the server must pass the session state from OpenSSL to
the kernel. OpenSSL does not implement an API to extract this data.
The current approach is to locally redefine some internal structs.
This is obviously fragile. On top of that, it requires OpenSSL 1.0,
as 1.1+ purposely hides even more struct types.

Signed-off-by: Willem de Bruijn <willemb@google.com>