ktls: rename and document SSL_generate_key_block output
Rename rx/tx to client/server to better match reality.

The test defines a struct with the output from SSL_generate_key_block.
The current layout uses _rx/_tx suffixes and works correctly when
passing to TLS_RX/TLS_TX on the server.

But the struct is actually defined in terms of client/server, which
means that this order has to be reversed for client sockets (opened
with SSL_connect).

Rename the struct fields to better match their contents and document
this subtle difference between RFC and KTLS terminology.

Signed-off-by: Willem de Bruijn <willemb@google.com>